+++ test/core/login/login.test.ts
@@ -533,4 +533,20 @@ describe('duress', function () {
     expect(await duressAccount.checkPassword(fakeUser.password)).equals(true)
     expect(await duressAccount.checkPassword('wrong password')).equals(false)
   })
+
+  it('change password', async function () {
+    const world = await makeFakeEdgeWorld([fakeUser], quiet)
+    const context = await world.makeEdgeContext(contextOptions)
+    const account = await context.loginWithPIN(fakeUser.username, fakeUser.pin)
+    await account.changePin({ pin: '0000', forDuressAccount: true })
+    const duressAccount = await context.loginWithPIN(fakeUser.username, '0000')
+
+    await duressAccount.changePassword('foobar')
+
+    expect(await duressAccount.checkPassword(fakeUser.password)).equals(true)
+    expect(await duressAccount.checkPassword('wrong password')).equals(false)
+
+    expect(await account.checkPassword(fakeUser.password)).equals(true)
+    expect(await account.checkPassword('wrong password')).equals(false)
+  })
 })